requirements.txt
# TornadoAI - WIN5特化予想サービス
anthropic>=0.40.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
lxml>=5.0.0
flask>=3.0.0
flask-cors>=6.0.0
gunicorn>=21.0.0
redis>=5.0.0
supabase>=2.0.0
